Create README.md
I am a life & fitness coach, and a Counselor. I have worked on myself to be the coach that I am today; my mission is to help as many people as I can reach their goals and ground themselves. 
During my career, I have worked in training and leading people in many aspects, on the other hand, I have managed many projects that catered for community service in the local society.